LayoutTests/fast/mediastream/MediaStreamTrack-getSources-expected.txt

Tests MediaStreamTrack::getSources.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS MediaStreamTrack.getSources(gotSources1); did not throw exception.
PASS gotSources1 was called.
PASS sources.length > 0 is true
PASS gotStream was called.
PASS MediaStreamTrack.getSources(gotSources2); did not throw exception.
PASS gotSources2 was called.
PASS sources.length > 0 is true
PASS sources[0].id === previousId is true
PASS sources[0].label.length > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
